requirements.txt
# Core dependencies for prodscope
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-core>=0.3.68
langgraph>=0.2.0

# LLM integrations
langchain-google-genai>=2.1.9
langchain-openai>=0.3.30
google-genai>=0.8.0  # For native Gemini search capabilities

# Google Cloud Vertex AI (Global access solution)
google-cloud-aiplatform>=1.45.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# For data analysis
scikit-learn>=1.3.0

# For report generation
jinja2>=3.1.0
markdown>=3.5.0

# Development tools
pytest>=7.4.0
python-dotenv>=1.0.0